salvando

+++ app/Http/Controllers/Api/V2/TaskController.php
@@ -6,16 +6,23 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
 use App\Models\Task;
+use Illuminate\Http\Request;
 
 class TaskController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
+        if ($request->user()->cannot('viewAny', Task::class)) {
+            abort(403);
+        }
+
         return request()->user()
-            ->tasks()->toResourceCollection();
+            ->tasks()
+            ->get()
+            ->toResourceCollection();
     }
 
     /**
@@ -25,6 +32,10 @@ class TaskController extends Controller
     {
         // $task = Task::create($request->validated() + ['user_id' => $request->user()->id]);
 
+        if ($request->user()->cannot('create', Task::class)) {
+            abort(403);
+        }
+
         $task = $request->user()->tasks()->create($request->validated());
 
         return $task->toResource();
@@ -33,11 +44,15 @@ class TaskController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Task $task)
+    public function show(Task $task, Request $request)
     {
         // return new TaskResource($task);
 
         // return TaskResource::make($task);
+
+        if ($request->user()->cannot('view', $task)) {
+            abort(403);
+        }
 
         return $task->toResource();
     }
@@ -47,6 +62,10 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        if ($request->user()->cannot('update', $task)) {
+            abort(403);
+        }
+
         $task->update($request->validated());
 
         return $task->toResource();
@@ -55,8 +74,12 @@ class TaskController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Task $task)
+    public function destroy(Task $task, Request $request)
     {
+        if ($request->user()->cannot('delete', $task)) {
+            abort(403);
+        }
+
         $task->delete();
 
         return response()->noContent();
